+++ api/api_athena/src/modules/minbar_predictor/ma_pred/ma_cal/ma_lwma.h
@@ -32,13 +32,13 @@ public:
         return _ins;
     }
 
-    real32 compLatestMA(std::vector<real32>& arr, int lookback, size_t i)
+    real64 compLatestMA(std::vector<real64>& arr, int lookback, size_t i)
     {
         if (i < lookback-1) {
             return NAN;
         }
 
-        real32 s = 0.f;
+        real64 s = 0.f;
         int w = 1;
         float W = (1. + lookback) * lookback*0.5;
         for (int m = i - lookback + 1; m <= i; m++) {
@@ -47,7 +47,7 @@ public:
         return s/W;
     }
 
-    void compAllMA(std::vector<real32>& arr, int lookback, std::vector<real32>& out)
+    void compAllMA(std::vector<real64>& arr, int lookback, std::vector<real64>& out)
     {
         out.resize(arr.size());
         for (size_t i = 0; i < out.size(); i++) {
